:ok_hand:refactored code into class and functional components for easier reading
